Bug fixes for review and trusted seller, plan cards for manage sales page
