+++ src/modules/dashboard/dashboard.controller.js
@@ -105,25 +105,6 @@ export const analyticsCount = async (req, res) => {
   }
 };
 
-import { getOemPlansWithRawSQL } from "./oemService.js";
-import { prepareOemPlansData } from "../../helpers/oemHelper.js";
-
-export const fetchPlansInfo = async (req, res) => {
-  const { fetch_plans_info } = req.query;
-  const vendor_id = req.user.vendor_id; // fixed !!
-  if (fetch_plans_info == 1) {
-    try {
-      const rawPlans = await getOemPlansWithRawSQL(vendor_id);
-      const preparedPlans = prepareOemPlansData(rawPlans);
-      return res.json({ plans: preparedPlans });
-    } catch (error) {
-      console.error(error);
-      return res.status(500).json({ error: "Failed to fetch plans data" });
-    }
-  }
-
-  return res.status(400).json({ message: "fetch_plans_info is not set to 1" });
-};
 // *******************************************************************************
 export const getDashboardStats = async (req, res) => {
   try {
