+++ src/modules/sales/sales.route.js
@@ -1,8 +1,9 @@
 import express from "express";
-import { planSubscribeRequest } from "./sales.controller.js";
+import { planSubscribeRequest, fetchPlansInfo } from "./sales.controller.js";
 
 const router = express.Router();
 
 router.post("/plan-subscribe-request", planSubscribeRequest);
+router.get("/oemData", fetchPlansInfo);
 
 export default router;
